test: add perf test helper for counting rows read
(cherry picked from commit c73d9fb7838178bb3d01db61ef3c76c38157b0dd)
